DjangoBackend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
gunicorn>=21.2
python-decouple>=3.8
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.3
drf-spectacular>=0.26.0